Correct type mapping of device index

+++ aten-scala/core/src/main/java/aten/TensorOptions.java
@@ -39,8 +39,8 @@ public class TensorOptions {
   public native boolean isCPU();
   public native boolean isCuda();
   public native TensorOptions cuda();
-  public native TensorOptions cuda_index(int device);
-  public native short deviceIndex();
+  public native TensorOptions cuda_index(short device);
+  public native int deviceIndex();
   public native byte scalarTypeByte();
 
   public static TensorOptions fromScalarType(byte i) {
